requirements.txt
pyodbc
pandas
azure-identity
python-dotenv
azure-storage-blob
pyarrow
pandas
pillow